Improve Fake location Handling
Because community toolkit toast alerts fault on Windows and the toolkit is not yet supported on .NET 9 avoid using it where possible. Switch to using ShowAppSnackBarAsync as a notification of a location change.

Note that picking a location from a map is not working on Windows but again, is not supported.