Improve Local/Remote Icon Display in Bill List
The remote status of bills is now shown correctly at startup as long as cloud access is allowed ( "Allow Cloud Backup" in Settings is enabled).

The icon logic is moved into XAML and the local Icon always shows, even if al it shows is that there's no local file. The remote icon doesn't show if cloud access is turned off.