Clean Up Unrecognized Licenses
Finalize the purchase of and licene passed to  Billing.GetInAppBillingPurchaseAsync, on the principle that it iwas probably created by a problem during the purchase process and the client already paid for it, so backing it out would be a pain.

Rename billing to inAppBilling so as to prevent confusion with Billing and do a bit of general cleanup.