Image Present Indication and Undelete Image
Lists of bills now include an icon to indicate whether or not each bill has a stored image associated with it.

The Image page now permits a deleted image to be recovered by usingthe "Undelete" menu selection (or using the Undelete button shown if there is no current image).

There's a new HasDeletedImage property of meals, updated pretty much whenever HasImage is.