Improve Properties Page User Interface
The font size has been increased slightly and when an entry field is selected it is enlarged further to make data entry easier.

When an entry field gets focus it is enlarged a little to make it clear what is being edited.

If you press enter after entering a value the keyboard will be dismissed and focus will be removed from the field.