Increase the Data Entry Wait Time
DivisiBill waits for the user to stop providing input during data entry on the properties page before attempting to update other items based on what's entered. This delay has been increased, as before it is possible to signal 'done' by pressing the enter key (which may be labeled 'done' or simply have a symbol on it.