Suppress CsWinRT1028
"Class should be marked partial" suppressed. It normally results in "Class 'Whatever' implements WinRT interfaces but it or a parent type isn't marked partial.  Type and any parent types should be marked partial for trimming and AOT compatibility if passed across the WinRT ABI.

But we don't use trimming and AOT on WinRT.