Initial Popup For license Check
When DivisiBill starts it checks a web service for licensing, if this check does not quickly succeed it would display a series of questions. Now it shows a popup while it retries errors and/or waits for responses. the popup closes automatically if a good response is received or you elect to continue without waiting for licenses.

It is implemented by introducing a new `CheckWebPage` popup using CommunityToolkit for .NET MAUI. `App.xaml.cs` to handle asynchronous web service checks displaying the popup on failure or timeout and passing a`Task<HttpStatusCode> representing the web services 'version' call.

`CheckWebPage.xaml` sets the popup size so it works of Windows and Android.

Most of the work is in `CheckWebPageViewModel` for data binding and the popup layout is defined  in CheckWebPage.xaml`. Because the popup dismisses itself when things work the  timeout in the initial `OrDelay` call method from 3000ms to 1000ms in `Utilities.cs` because the dialog goes away automatically.

The `CheckWebPageViewModel` constructor accepts parameters directly passed in via the 'CheckWebPage' constructor.