Improve Format of "Telemetry" Popup
During initialization DivisiBill may ask the user whether they wish to send telemetry in the event of a failure. The appearance of this dialog has been improved.

Updated QuestionPage.xaml to implement a toolkit:Popup with a transparent background, rounded border, and improved layout. Adjusted padding, background color, and spacing for better visual consistency. CanBeDismissedByTappingOutsideOfPopup is set to False to guarantee to call Button_Clicked in order to set up a result object.