Change Cloud Timeout to 30s
If the web service is not responsive we now allow 30s before giving up instead of waiting 100s as we did before.

Updated the `HttpClient` initialization in the `CallWs` class to include a `Timeout` property set to 30 seconds. This change improves request handling by preventing hanging requests and enhancing application reliability.