Stop Counting When In Background
The dialog that pops up when failing to call a web service has a countdown to retry and a count of seconds since the last try. With this change the count is not updated when the app is in the background.

Enhance clarity of debug messages in `App.xaml.cs` by using `==` for condition checks.
Add a new debug message in `CallWs.cs` to indicate when a successful response has no data.
Adjust status messages in `CheckWebPageViewModel.cs` to accurately reflect the application's running state, especially when paused, and correct the timing of the elapsed timer.