Check for Image File Presence
When restoring a bill from an archive or downloading it from the cloud the application now checks for the presence of an image or a deleted image.

Introduced the `CheckImageFiles` method in the `MealSummary` class to centralize the logic for checking image file existence, replacing the previous methods `DetermineHasImage` and `DetermineHasDeletedImage`.

Updated various parts of the `Meal` and `MealSummary` classes to call this new method, ensuring consistent image file status updates. Removed redundant calls to image file checks from the `Meal` class and the `ImageViewModel`, improving code organization and maintainability.

The `Archive` class now also utilizes `CheckImageFiles` to initialize image file status for restored meals.