Change Sharing now includes even Sharing
On the items page you may change the current sharing of an item by tapping in the left margin or by selecting the item, using the menu for the page and selecting 'Change'.

Updated the LineItem class to include a new enumeration `SharingType` with values `Even`, `Uneven`, and `None`. Introduced the `GetSharingType` method to determine sharing type based on sharer and share counts. Refactored the `SharesList` property to utilize a new `GetMaxDiner` method, which efficiently retrieves the highest DinerID with positive shares. Additionally, modified the MealViewModel to use the new `GetSharingType` method for determining changeType, enhancing clarity and maintainability.