Show Meal List Download Icons
When cloud access was turned on as a result of a request to show remote bills the remote/local status icons on each bill were not shown. This has been corrected.

The basic fix was to move a number of properties from CloudViewModel to SettingsViewModel so as to more easily handle dynamic updates.