Refactor MapPage Use
Navigation to and from MapPage instances now uses Shell Navigation and a new class (MapSettings) to pass input and receive responses from a MapPage.

There are no functional changes, the user view of mapping is unchanged. In the course of this I discovered a bug in the community toolkit Maps class (you can't seem to navigate away from it without a fault) but it's going away anyway so not much point getting excited.