Show a Few Bills Per Group
When Grouping bills by venue on a bill list page only the first few bills for each venue are shown.

To see all the bills for a venue you can now expand the venue to see the most recent bills or if there are a lot tap the "all bills" button to open a new page and see all the bills for that venue.

Summary of changes:

Add CountText and FirstMealSummaries properties

Updated MealSummaryGroup to include CountText for formatted meal counts and FirstMealSummaries to limit displayed meals. Modified MealSummaries_CollectionChanged to regenerate FirstMealSummaries on changes. Updated MealListPage.xaml to bind to CountText and FirstMealSummaries, ensuring only the first maxMeals are shown.

Reformatted BoxView attributes to make it as obvious as possible that it generates a simple ruled line.

Updated App.xaml.cs to add new shell navigation methods
for relative and absolute navigation with query parameters.
Implemented IQueryAttributable in MealListViewModel.cs
to handle query parameters and added filtering and sorting
capabilities.

- Introduced `IsStringNotNullOrWhiteSpaceConverter` and
  `IsStringNullOrWhiteSpaceConverter` in App.xaml for data binding.
- Refactored `VenueName` in `MealListViewModel.cs` to use
  `[ObservableProperty]` and renamed it FilteredVenueName.
- Updated `Label` bindings in `MealListPage.xaml` to utilize
  `MultiBinding` for dynamic visibility based on `IsGrouped`
  and `FilteredVenueName` conditions.
- Added a new `Label` to display `FilteredVenueName` with formatting,
  ensuring visibility only when `FilteredVenueName` is not null or whitespace.