Improved Bill List Layout
The individual entries in a bill list are laid out slightly more neatly, especially when the list is grouped by venue name.

The MealList DataTemplate is simplified considerably.

Uses just a single grid in place of the previous nested grids.

Uses a single DataTemplate because the SwipeView on Windows no longer crashes.

The 'model' namespace is renamed 'models' for consistency.

The 'All' value converter is moved to App.XAML.