Improve meal List User Interface Responsiveness
You can tap in the center of an item to select or deselect it but for the past couple of releases that tap would be mistaken for a swipe initiation. This no longer happens, swipe actions may only be initiated from right or left of center.

The fix was to overlay a transparent button so that the SwipeView never even sees a swipe. The side effect is that we no longer need fancy dancing using timers on Android.

Removed the SelectedMealSummaryChangeTime field and related methods from MealListViewModel, simplifying the selection logic. Restructured the XAML in MealListPage to introduce a new Grid layout to allow for a transparent button overlaying the SwipeView.