Add a "Scroll to Current" Bill Command
The bill list pages now have a command to scroll the current bill or venue into view.

Simplified the `OnNavigatedTo` method in `MealListViewModel` to call a new `ShowCurrent` method. This method determines the target index of the current meal and scrolls to it. Introduced `ShowCurrentCommand` for UI interaction and added a "Scroll To Current" button in the `MealListPage.xaml` toolbar.