Always Allow Diner Undelete
Rarely, a stored bill would have non-contiguous DinerIDs, for example 1 and 6. In this case the dinerID would be shown on the assignment button which was unnecessary and confusing for the user.

What's more in the example above, if you were to delete and undelete Diner 6, 6 the program would crash. This has been corrected and 6 would simply be renumbered 2 in this case. Since the DinerID number is rarely user visible anyway the user does not see a significant difference.