Add WinUI checks for map availability
The community toolkit Maps control for Windows is going away (it's not in version 3.0.2) because Bing Maps is going away, so do not use Maps on Windows.

Introduce conditional checks for Windows UI (WinUI) to display a message when the map feature is unavailable. Updated `SetFakeLocation` in `SettingsViewModel`, `ShowMap` in `VenueEditViewModel`, and `OnShowMap` in `VenueListPage`. Modified `MauiProgram` to conditionally use `UseMauiMaps` based on the platform not being Windows.