Create a Sentry Event for Unexpected navigation
There's no navigation UI on the splash page yet Play Store testing seems to do navigate to other pages before the splash page has completed. This is just to detect if that's happening still. It does this by:

Adding PrematureNavigationTitle for Sentry reporting (so we can test for it and send unconditionally if it happens).

Introduce a new constant `PrematureNavigationTitle` in the
`SentryEventProcessor` class to categorize navigation errors.
Update Sentry event logic to include this title alongside
`UserFeedbackTitle`. Enhance error reporting in the
`SplashPage` class by sending detailed messages to Sentry
when navigation is canceled due to non-popup sources.