Correct 'Save' Action on Properties Page
Previously it was not correctly saving any image file associated with the bill or entering the newly saved bill into the bill list. These errors have been corrected.

The major problem was that `LoadFromApp` would reuse a local MealSummary object if one existed, which is fine for the normal case but not for the 'Save' case. In detail:

- Updated LoadFromApp to accept a boolean parameter `tryExistingSummary` for conditional loading of existing meal summaries.
- Modified SaveSnapshotAsync to load a meal with an independent summary by passing `tryExistingSummary: false`.
- Changed the `IsLocal` property of the `Summary` object to `true` in SaveSnapshotAsync, indicating the meal is now local.
- Corrected functionality to copy existing images to the new location and update the summary accordingly.