Make properties Shell Selection Navigable
The "Properties" selection now allows use of the hamburger menu to navigate away from it. The same page shown by tapping the venue name on the "Items" or "Totals" pages returns only to those pages, as before.

This was mode possible by noting that the same page may be in the XAML and declared using Routing.RegisterRoute.

The properties page now needs to set Shell.FlyoutBehavior dynamically.