Update Bill List When Location Changes
When the app location changes by a significant amount (more than 20m currently) update the distances to venues reported in the bill list. This was being don in the list sorted by distance, but not otherwise, that has been corrected.

Similar "at least 20m" logic has been applied to the case of fake locations for testing, so multiple calls to GetMyLocation are a no-op.

MealSummary.Distance is now Observeable.

Added a Utilities.GetDistanceBetween function for clarity.