Improve Venue Edit Page
The app will no longer display a count of bills using the venue being edited if the venue name is not changed.

Refactor logic to check if the venue name has changed before prompting the user for confirmation. If the name has changed, the user is asked for confirmation before saving changes. If the name remains the same, changes are saved without additional prompts,.