Change the Alternate Web Service Debug Option
Instead of a setting that determines whether to use the alternate web service next time the app is run there's now a setting to determine whether to ask about using it.

The CallWs.SelectWs function has been simplified and renamed SelectAlternateWs. It is now only called out of InitializeApp if it is needed.

Settings.UseAlternateWs has been renamed AskAboutUsingAlternateWs, which accounts for most of the changed lines.

A couple of unnecessary calls on App.HandleActivityChanges have been eliminated.