Handle Requests to Display Payments on an Item
Requesting a payment summary from a specific item now presumes you intend it for the first sharer of that item unless you are filtering for a particular participant in which case they are assumed to be the target.

Introduced a new `DisplayPaymentsForLineItem` method in the
`MealViewModel` to handle payments at the `LineItem` level.
This method determines the associated `PersonCost` and
invokes the `DisplayPayments` method accordingly.

Updated `LineItemsPage.xaml` to replace the double-tap
gesture for `DisplayPaymentsCommand` with
`DisplayPaymentsForLineItemCommand`, using `SelectedLineItem`
as the `CommandParameter`.