Speed Up Page Loading
Several pages (Totals, People, Items, People Lists and venue Lists) have been optimized to speed them up a little. There only functional changes are minor  formatting enhancements.

The bug which required that SwipeView not be used on Windows (https://github.com/dotnet/maui/issues/18530) is fixed, so just put the required transparent overlay buttons in the "WithSwipe" format on Windows only and use that.  One key technique is to use an ImageButton for both Windows and Android because it has no lower limit on height and is not susceptible to the weird event ordering that plagues other controls.