Item Insert User Interface Changes
When inserting an item on the items page it will be inserted after the selected item. If no item is currently selected the new item will be added at the end of the list. The new item will be selected once it has been added.

If you want to inset an item before the first one, select the first one, insert the item, then delete the first one. The selection will switch to the new item and tapping "Undelete" will restore the former first item after the new one.

Refactored `LineItem.nextItemNumber` into a property `NextItemNumber` to improve encapsulation and consistency. Introduced `TrySetNextItemNumberFromName` and `DefaultItemNumber` methods to simplify and isolate logic for updating item numbers. Updated `OnItemNameChanged` to use these new methods.

Replaced all occurrences of `nextItemNumber` in `Meal` with the new `NextItemNumber` property and updated `savedNextItemNumber` to match the new type (basically using int instead of uint). Removed redundant resets of `nextItemNumber`.

Added a new `InsertAfter` extension method in `Utilities` to enhance list manipulation, providing a counterpart to `InsertBefore`. Updated `MealViewModel.AddItem` to use `InsertAfter` for inserting items after the selected item or at the end if none is selected.