Allow "Retry" Option When Waiting for Web Response
The initial web request waits for 30s between failed calls, it is now possible to proceed with the next call without waiting by tapping the "Retry" button. The existing "Give Up" button is unchanged.

This is managed with a new RequestRetry command and supporting logic in the view model. The UI updates to show the button only during countdowns, and layout and debug messages are improved for clarity (in particular, the popup has constant width).