Show Simple Coupon Amount in UI Instead of Modified value
A bill with coupons applied after tax would show an incorrect value for the sum of the coupons (it was actually showing what the coupon would be worth before tax). The actual coupon face value is once again shown.

Made GetRawCouponAmount public and exposed it via MealViewModel. Updated LineItemsPage and TotalsPage to display RawCouponAmount, showing the sum of all coupons rather than the calculated-as-if-it-were-before-tax coupon amount.